Using git again, see description for details
Previously this was using unity collab for source control as it launched during this project, however 3 months in git still seems superior so this project returns to git with this commit.
NOTE: This build is unstable, this push is only to change source control, and happens during the change.

The current change in development is a network design change. Previous design aimed to only send messages when players issued movement changes, however this is very complex, and it seems as if this might actually increase the amount of packages sent instead of decreasing it. The new design will send X packages every second.